Convert to spaces

+++ src/systems/system.h
@@ -7,19 +7,19 @@
 class Entity;
 
 class System : public MessageHandler {
-	public:
-		virtual ~System() = default;
+  public:
+    virtual ~System() = default;
 
-		virtual void attachEntity(Entity* entity) = 0;
-		virtual void detachEntity(Entity* entity) {
-			for (unsigned int i = 0; i < entities.size(); i++) {
-				if (entities[i] == entity) {
-					entities.erase(entities.begin() + i);
-					break;
-				}
-			}
-		};
-		virtual void loop() = 0;
+    virtual void attachEntity(Entity* entity) = 0;
+    virtual void detachEntity(Entity* entity) {
+      for (size_t i = 0; i < entities.size(); i++) {
+        if (entities[i] == entity) {
+          entities.erase(entities.begin() + i);
+          break;
+        }
+      }
+    };
+    virtual void loop() = 0;
 
-		std::vector<Entity*> entities;
+    std::vector<Entity*> entities;
 };
